pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Agentic RAG chatbot for EU AI Act + GDPR compliance (LangGraph + hybrid vector/graph RAG)"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "langgraph>=0.2.50",
    "langchain-core>=0.3.20",
    "langchain-community>=0.3.5",
    "chromadb>=0.5.20",
    "sentence-transformers>=3.3.0",
    "networkx>=3.4",
    "pypdf>=5.1.0",
    "pymupdf>=1.24.0",
    "pydantic>=2.9.0",
    "streamlit>=1.40.0",
    "ollama>=0.4.4",
    "duckduckgo-search>=6.3.0",
    "httpx>=0.27.0",
    "tqdm>=4.66.0",
    "rich>=13.9.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_rag"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
